simple_graphrag/pyproject.toml
[project]
name = "simple-graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "networkx>=3.6.1",
    "openai>=2.13.0",
    "pyyaml>=6.0.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
